requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2==2.9.6
drf-spectacular==0.26.3
Pillow==10.0.0
